pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-eval-engine"
version = "0.1.0"
description = "Starter PyTorch project with a simple training pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "you@example.com"}
]
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "numpy>=1.24",
  "tqdm>=4.64",
  "pydantic>=2.5",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.13",
  "mypy>=1.8"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
